+++ src/ATK/Tools/query.py
@@ -28,18 +28,25 @@ def query(kind: str, **arguments) -> DataSet:
     if targets is RETURNS.NULL:
         raise ValueError("Query received no targets.")
 
+    if targets is RETURNS.EXCEPTION:
+        raise Exception("Failed to generate requested targets, this is likely due to a Vizier fault.")
+
     # targets may return exception if Vizier is down
     if any(target is RETURNS.EXCEPTION for target in targets):
         warnings.warn("Failed to generate requested targets, this is likely due to a Vizier fault.")
 
-        structure = DataSet[kind](kind=kind, survey=arguments.get("survey", None), targets=None, radius=arguments.get("radius", None), exception=True)
+        structure = DataSet[kind](
+            kind=kind, survey=arguments.get("survey", None), targets=None, radius=arguments.get("radius", None), exception=True
+        )
         return structure
 
     # disable proper motion correction
     if arguments.get("disable_correction", False):
         for target in targets:
             initial_coords = target.initial_coords
-            target.initial_coords = SkyCoord(ra=initial_coords.ra, dec=initial_coords.dec, frame=initial_coords.frame, obstime=initial_coords.obstime)
+            target.initial_coords = SkyCoord(
+                ra=initial_coords.ra, dec=initial_coords.dec, frame=initial_coords.frame, obstime=initial_coords.obstime
+            )
             target.coords = initial_coords
             target.identifier = None
             target.survey = None
@@ -49,6 +56,9 @@ def query(kind: str, **arguments) -> DataSet:
 
     # attach cache key
     structure._cache_key = make_cache_key(kind, targets, arguments)
+
+    if structure.exception:
+        return structure
 
     # save if path provided
     if arguments.get("path"):